Added one bool to check all loaded data for errors.
The DBLS (Database Load Struct) returned by the function "DBLoad()" now
contains a boolean variable "noErrors". It will be set true if the
database did not reject any connection attempted and false if any
database connections failed.

noErrors == true does not imply all vectors have data, only that if
there was relevant data they will be in the vectors. Meaning right now
the characters vector is still going to be empty seeing as this database
hasn't been used to play the game yet.