Code So Far
This is my code so far.  It is a mere skeleton of what it needs to be.
I have commented in my code what needs to go where.  All windows are
connected to each other as they're supposed to be but I'm having a
problem with the WizardSpecialization.  For some reason It gives me an
unidentified external error.